fix(api): remove duplicate HttpGet on canvas endpoint, add PUT and XML docs
- Duplicate [HttpGet] attribute on canvas endpoint caused ASP0023 route conflict
- Added PutCanvas with [HttpPut] to persist canvas HTML via REST (CS1591 fix)
- Fixed missing $canvasHtml SQLite parameter bind in SaveConversationAsync
- Extracted StoreFixture to shared file so CanvasPersistenceTests can reference it

Closes #412
